Move definitions from wincon to wincontypes
