+++ Cargo.toml
@@ -331,6 +331,7 @@ winbase = []
 wincodec = []
 wincodecsdk = []
 wincon = []
+wincontypes = []
 wincred = []
 wincrypt = []
 windowsceip = []
