+++ src/um/consoleapi.rs
@@ -6,7 +6,8 @@
 // except according to those terms.
 //! ApiSet Contract for api-ms-win-core-console-l1
 use shared::minwindef::{BOOL, DWORD, LPDWORD, LPVOID, UINT};
-use um::wincon::{COORD, HPCON, PCONSOLE_READCONSOLE_CONTROL, PHANDLER_ROUTINE, PINPUT_RECORD};
+use um::wincon::{PCONSOLE_READCONSOLE_CONTROL, PHANDLER_ROUTINE};
+use um::wincontypes::{COORD, HPCON, PINPUT_RECORD};
 use um::winnt::{HANDLE, HRESULT, VOID};
 extern "system" {
     pub fn AllocConsole() -> BOOL;
