+++ src/um/mod.rs
@@ -238,6 +238,7 @@ pub mod gl;
 #[cfg(feature = "wincodec")] pub mod wincodec;
 #[cfg(feature = "wincodecsdk")] pub mod wincodecsdk;
 #[cfg(feature = "wincon")] pub mod wincon;
+#[cfg(feature = "wincontypes")] pub mod wincontypes;
 #[cfg(feature = "wincred")] pub mod wincred;
 #[cfg(feature = "wincrypt")] pub mod wincrypt;
 #[cfg(feature = "windowsceip")] pub mod windowsceip;
